requirements.txt
beautifulsoup4==4.11.1
multipledispatch==0.6.0
mysql_connector_repackaged==0.3.1
pandas==1.3.5
pytest==6.2.5
requests==2.27.0
selenium==4.1.5
google-cloud-bigquery-storage==2.13.2
google-cloud-bigquery==3.1.0
pandas-gbq==0.17.5